ccloud: retry cifs setup and keep failed vservers
Keep share servers on error:
either if they have a sec services config
or if the 'keep_share_server_on_failure' option is turned on.

But disable the lifs, since the neutron ports are also deleted
(for non cloud admins).
They can be enabled manually on the filer for troubleshooting.
But keep in mind, that the IP may have been re-used.
Fallback to setup cifs without ldaps and session signing after 3 failed attempts.
ensure: don't re-apply cifs security settings to allow manual override

(cherry picked from commit a0a04c18113bdf8fff6d4d086f58ecf6d857f9be)

+++ manila/share/drivers/netapp/dataontap/cluster_mode/lib_multi_svm.py
@@ -21,7 +21,7 @@ as needed to provision shares.
 """
 import re
 
-from manila.share.drivers.netapp.dataontap.cluster_mode.lib_base import Backup
+from oslo_config import cfg
 from oslo_log import log
 from oslo_serialization import jsonutils
 from oslo_utils import excutils
@@ -36,12 +36,22 @@ from manila.share.drivers.netapp.dataontap.client import client_cmode
 from manila.share.drivers.netapp.dataontap.client import client_cmode_rest
 from manila.share.drivers.netapp.dataontap.cluster_mode import data_motion
 from manila.share.drivers.netapp.dataontap.cluster_mode import lib_base
+from manila.share.drivers.netapp.dataontap.cluster_mode.lib_base import Backup
 from manila.share.drivers.netapp import utils as na_utils
 from manila.share import share_types
 from manila.share import utils as share_utils
 from manila import utils
 
 
+lib_multi_svm_opts = [
+    cfg.BoolOpt('keep_share_server_on_failure',
+                default=False,
+                help='Whether share servers will '
+                     'be deleted on failures.'),
+]
+
+CONF = cfg.CONF
+CONF.register_opts(lib_multi_svm_opts)
 LOG = log.getLogger(__name__)
 SUPPORTED_NETWORK_TYPES = (None, 'flat', 'vlan')
 SEGMENTED_NETWORK_TYPES = ('vlan',)
@@ -367,9 +377,18 @@ class NetAppCmodeMultiSVMFileStorageLibrary(
                     LOG.warning("Failed to configure Vserver.")
                     # NOTE(dviroel): At this point, the lock was already
                     # acquired by the caller of _create_vserver.
-                    self._delete_vserver(vserver_name,
-                                         security_services=security_services,
-                                         needs_lock=False)
+                    # NOTE(carthaca): keep debris for analysis in debug
+                    if not (CONF.keep_share_server_on_failure or security_services):  # noqa: E501
+                        self._delete_vserver(vserver_name,
+                                             security_services=security_services,  # noqa: E501
+                                             needs_lock=False)
+                    # NOTE(carthaca): lifs need to be turned off to not provoke
+                    # duplicate IPs
+                    else:
+                        lifs = vserver_client.get_network_interfaces()
+                        for lif in lifs:
+                            self._client.disable_network_interface(
+                                vserver_name, lif['interface-name'])
 
     def _setup_network_for_vserver(self, vserver_name, vserver_client,
                                    network_info, ipspace_name,
